added custom metric RTP to QA app
